+++ physicsnemo/experimental/models/dit/dit.py
@@ -18,7 +18,7 @@ from typing import Tuple, Union, Optional, Literal
 import torch
 import torch.nn as nn
 
-from physicsnemo.models.utils import PatchEmbed2D, PatchRecovery2D
+from physicsnemo.models.utils import PatchEmbed2D
 from physicsnemo.models.diffusion import PositionalEmbedding, Linear
 from dataclasses import dataclass
 from physicsnemo.models.meta import ModelMetaData
@@ -32,10 +32,13 @@ class MetaData(ModelMetaData):
     jit: bool = False
     cuda_graphs: bool = False
     amp_cpu: bool = False
-    amp_gpu: bool = False
+    amp_gpu: bool = True
     torch_fx: bool = False
-    bf16: bool = False
+    # Data type
+    bf16: bool = True
+    # Inference
     onnx: bool = False
+    # Physics informed
     func_torch: bool = False
     auto_grad: bool = False
 
@@ -73,10 +76,8 @@ class DiT(Module):
         If 'apex', uses FusedLayerNorm from apex. If 'torch', uses LayerNorm from torch.nn. Defaults to 'apex'.
     condition_dim (int, optional):
         Dimensionality of conditioning. If None, the model is unconditional. Defaults to None.
-    embedding_type (str, optional):
-        The type of positional embedding ('sin-cos' or 'learnable'). Defaults to 'sin-cos'.
-    pos_embedding_dim (int, optional):
-        The dimensionality of the positional embedding. Defaults to 1.
+    dit_initialization (bool, optional):
+        If True, applies the DiT specific initialization. Defaults to True.
 
     Forward
     -------
@@ -126,7 +127,7 @@ class DiT(Module):
         attention_backbone: Literal["timm", "transformer_engine"] = "transformer_engine",
         layernorm_backbone: Literal["apex", "torch"] = "torch",
         condition_dim: Optional[int] = None,
-        pos_embedding_dim: int = 1,
+        dit_initialization: Optional[int] = True,
     ):
         """
         Initializes the DiT model.
@@ -145,16 +146,17 @@ class DiT(Module):
         self.x_embedder = PatchEmbed2D(
             self.input_size,
             self.patch_size,
-            in_channels + pos_embedding_dim,
+            in_channels,
             hidden_size,
         )
-        self.t_embedder = PositionalEmbedding(hidden_size)
+        self.t_embedder = PositionalEmbedding(hidden_size, amp_mode=self.meta.amp_gpu, learnable=True)
         init_zero = dict(init_mode="kaiming_uniform", init_weight=0, init_bias=0)
         self.cond_embedder = (
             Linear(
                 in_features=condition_dim,
                 out_features=hidden_size,
                 bias=False,
+                amp_mode=self.meta.amp_gpu,
                 **init_zero,
             )
             if condition_dim
@@ -165,10 +167,8 @@ class DiT(Module):
         self.num_patches = self.h_patches * self.w_patches
 
         # Learnable positional embedding:
-        self.pos_embed = nn.Parameter(
-            torch.zeros(pos_embedding_dim, self.input_size[0], self.input_size[1]),
-            requires_grad=True,
-        )
+        self.pos_embed = nn.Parameter(torch.zeros(1, self.num_patches, hidden_size), requires_grad=True)
+
         self.blocks = nn.ModuleList(
             [
                 DiTBlock(
@@ -186,12 +186,38 @@ class DiT(Module):
             self.patch_size[0] * self.patch_size[1] * self.out_channels,
             layernorm_backbone,
         )
-        self.patch_recovery = PatchRecovery2D(
-            self.input_size,
-            self.patch_size,
-            self.patch_size[0] * self.patch_size[1] * self.out_channels,
-            self.out_channels,
-        )
+        if dit_initialization:
+            self.initialize_weights()
+
+
+    def initialize_weights(self):
+        # Apply a basic Xavier uniform initialization to all linear layers.
+        def _basic_init(module):
+            if isinstance(module, nn.Linear):
+                torch.nn.init.xavier_uniform_(module.weight)
+                if module.bias is not None:
+                    nn.init.constant_(module.bias, 0)
+        self.apply(_basic_init)
+
+        # Initialize the learnable positional embedding with a normal distribution.
+        nn.init.normal_(self.pos_embed, std=0.02)
+
+        # Initialize the patch embedding projection (a Conv2D layer).
+        w = self.x_embedder.proj.weight.data
+        nn.init.xavier_uniform_(w.view([w.shape[0], -1]))
+        if self.x_embedder.proj.bias is not None:
+            nn.init.constant_(self.x_embedder.proj.bias, 0)
+        
+        # DiT specific initializations
+        for block in self.blocks:
+            nn.init.constant_(block.adaptive_modulation[-1].weight, 0)
+            nn.init.constant_(block.adaptive_modulation[-1].bias, 0)
+
+        nn.init.constant_(self.proj_layer.adaptive_modulation[-1].weight, 0)
+        nn.init.constant_(self.proj_layer.adaptive_modulation[-1].bias, 0)
+        
+        nn.init.constant_(self.proj_layer.output_projection.weight, 0)
+        nn.init.constant_(self.proj_layer.output_projection.bias, 0)
 
     def forward(
         self, x: torch.Tensor, t: torch.Tensor, condition: Optional[torch.Tensor] = None
@@ -199,11 +225,9 @@ class DiT(Module):
         """
         Performs the forward pass of the DiT model.
         """
-        b, ch, h, w = x.shape
-        x = torch.cat([x, self.pos_embed.repeat(b, 1, 1, 1)], dim=1)
         x_emb = self.x_embedder(x)
         # (N, D, H//patch[0], W//patch[1])
-        x = x_emb.flatten(2).transpose(1, 2)
+        x = x_emb.flatten(2).transpose(1, 2) + self.pos_embed
         # (N, T, D) T = H//patch[0] * W//patch[1]
         t = self.t_embedder(t)  # (N, D)
 
@@ -221,8 +245,10 @@ class DiT(Module):
             c = t  # (N, D)
         for block in self.blocks:
             x = block(x, c)  # (N, T, D)
-        x = self.proj_layer(x, c)  # (N, T, D')
-        x = x.reshape(x.shape[0], x.shape[-1], self.h_patches, self.w_patches)
-        # (N, D', H//patch[0], W//patch[1])
-        x = self.patch_recovery(x)  # (N, out_channels, H, W)
+        x = self.proj_layer(x, c)  # (N, T, patch[0]*patch[1]*out_channels)
+        x = x.reshape(shape=(x.shape[0], self.h_patches, self.w_patches, 
+                             self.patch_size[0],self.patch_size[1], self.out_channels))
+        x = torch.einsum('nhwpqc->nchpwq', x) # (N, C, H//patch[0], W//patch[1], patch[0], patch[1])
+        x = x.reshape(shape=(x.shape[0], self.out_channels,
+                                 self.h_patches * self.patch_size[0], self.w_patches *self.patch_size[1]))
         return x