+++ physicsnemo/models/diffusion/layers.py
@@ -1038,7 +1038,8 @@ class PositionalEmbedding(torch.nn.Module):
         If True, the embedding considers the endpoint. By default False.
     amp_mode : bool, optional
         A boolean flag indicating whether mixed-precision (AMP) training is enabled. Defaults to False.
-
+    learnable : bool, optional
+        A boolean flag indicating whether learnable positional embedding is enabled. Defaults to False.
     """
 
     def __init__(
@@ -1047,12 +1048,20 @@ class PositionalEmbedding(torch.nn.Module):
         max_positions: int = 10000,
         endpoint: bool = False,
         amp_mode: bool = False,
+        learnable: bool = False,
     ):
         super().__init__()
         self.num_channels = num_channels
         self.max_positions = max_positions
         self.endpoint = endpoint
         self.amp_mode = amp_mode
+        self.learnable = learnable
+        if learnable:
+            self.mlp = torch.nn.Sequential(
+                torch.nn.Linear(num_channels, 2 * num_channels, bias=True),
+                torch.nn.SiLU(),
+                torch.nn.Linear(2 * num_channels, num_channels, bias=True),
+            )
 
     def forward(self, x):
         freqs = torch.arange(
@@ -1066,6 +1075,8 @@ class PositionalEmbedding(torch.nn.Module):
                 freqs = freqs.to(x.dtype)
         x = x.ger(freqs)
         x = torch.cat([x.cos(), x.sin()], dim=1)
+        if self.learnable:
+            x = self.mlp(x)
         return x